// src/lib/types/task.ts
export interface Task {
  id: string;
  doi_tuong: string;
  doi_tuong_id: string;
  loai_viec: string;
  dien_giai: string;
  nguoi_phu_trach: string;
  ten_viec: string;
  ghi_chu?: string;
  ngay_den_han: string;
  cap_do: string;
  trang_thai: string;
  ngay_hoan_thanh?: string;
  tg_tao?: string;
}